backend/requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.1
django-allauth==0.47.0
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-pytest==0.2.0
djangorestframework==3.13.1
drf-spectacular==0.21.2
drf-spectacular-sidecar==2022.2.7
drf-writable-nested==0.6.3
Faker==11.3.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
jsonschema==4.4.0
oauthlib==3.1.1
Pillow==9.0.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
